requirements.txt
pytest==3.0.5
scrapy==1.3.0
PyYAML==3.12
ipdb==0.10.1
beautifulsoup4==4.5.1
requests==2.12.4
